Credit_Score_Classification
The aim of this project is to develop a classification model for credit score assessment. Credit
score is a crucial metric used to evaluate the financial status of individuals or organizations. The
primary objective of this project is to leverage data analysis and machine learning techniques to
develop a classification model that assesses customers' credit risk.